Reverted to master state
