+++ packages/monorepo-builder/packages/testing/src/ComposerJsonRepositoriesUpdater.php
@@ -6,14 +6,41 @@ namespace Symplify\MonorepoBuilder\Testing;
 
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symplify\ComposerJsonManipulator\FileSystem\JsonFileManager;
+use Symplify\ComposerJsonManipulator\ValueObject\ComposerJson;
 use Symplify\ConsoleColorDiff\Console\Output\ConsoleDiffer;
 use Symplify\MonorepoBuilder\Package\PackageNamesProvider;
 use Symplify\MonorepoBuilder\Testing\ComposerJson\ComposerJsonSymlinker;
 use Symplify\MonorepoBuilder\Testing\PackageDependency\UsedPackagesResolver;
 use Symplify\SmartFileSystem\SmartFileInfo;
+use Symplify\SymplifyKernel\Exception\ShouldNotHappenException;
 
-final class ComposerJsonRepositoriesUpdater extends AbstractComposerJsonRepositoriesUpdater
+final class ComposerJsonRepositoriesUpdater
 {
+    /**
+     * @var PackageNamesProvider
+     */
+    private $packageNamesProvider;
+
+    /**
+     * @var JsonFileManager
+     */
+    private $jsonFileManager;
+
+    /**
+     * @var SymfonyStyle
+     */
+    private $symfonyStyle;
+
+    /**
+     * @var UsedPackagesResolver
+     */
+    private $usedPackagesResolver;
+
+    /**
+     * @var ConsoleDiffer
+     */
+    private $consoleDiffer;
+
     /**
      * @var ComposerJsonSymlinker
      */
@@ -27,29 +54,51 @@ final class ComposerJsonRepositoriesUpdater extends AbstractComposerJsonReposito
         ConsoleDiffer $consoleDiffer,
         ComposerJsonSymlinker $composerJsonSymlinker
     ) {
+        $this->packageNamesProvider = $packageNamesProvider;
+        $this->jsonFileManager = $jsonFileManager;
+        $this->symfonyStyle = $symfonyStyle;
+        $this->usedPackagesResolver = $usedPackagesResolver;
+        $this->consoleDiffer = $consoleDiffer;
         $this->composerJsonSymlinker = $composerJsonSymlinker;
-
-        parent::__construct(
-            $packageNamesProvider,
-            $jsonFileManager,
-            $symfonyStyle,
-            $usedPackagesResolver,
-            $consoleDiffer
-        );
     }
 
-    /**
-     * @param mixed[] $packageComposerJson
-     * @param string[] $packageNames
-     * @return mixed[]
-     */
-    public function decoratePackageComposerJson(array $packageComposerJson, array $packageNames, SmartFileInfo $rootComposerJsonFileInfo, ?bool $symlink): array
+    public function processPackage(SmartFileInfo $packageFileInfo, ComposerJson $rootComposerJson, ?bool $symlink): void
     {
-        return $this->composerJsonSymlinker->decoratePackageComposerJsonWithPackageSymlinks(
+        $packageComposerJson = $this->jsonFileManager->loadFromFileInfo($packageFileInfo);
+
+        $usedPackageNames = $this->usedPackagesResolver->resolveForPackage($packageComposerJson);
+        if ($usedPackageNames === []) {
+            $message = sprintf(
+                'Package "%s" does not use any mutual dependencies, so we skip it',
+                $packageFileInfo->getRelativeFilePathFromCwd()
+            );
+            $this->symfonyStyle->note($message);
+            return;
+        }
+
+        // possibly replace them all to cover recursive secondary dependencies
+        $packageNames = $this->packageNamesProvider->provide();
+
+        $oldComposerJsonContents = $packageFileInfo->getContents();
+
+        $rootComposerJsonFileInfo = $rootComposerJson->getFileInfo();
+        if ($rootComposerJsonFileInfo === null) {
+            throw new ShouldNotHappenException();
+        }
+
+        $packageComposerJson = $this->composerJsonSymlinker->decoratePackageComposerJsonWithPackageSymlinks(
             $packageComposerJson,
             $packageNames,
             $rootComposerJsonFileInfo,
             $symlink
         );
+
+        $newComposerJsonContents = $this->jsonFileManager->printJsonToFileInfo($packageComposerJson, $packageFileInfo);
+
+        $message = sprintf('File "%s" was updated', $packageFileInfo->getRelativeFilePathFromCwd());
+        $this->symfonyStyle->title($message);
+
+        $this->consoleDiffer->diffAndPrint($oldComposerJsonContents, $newComposerJsonContents);
+        $this->symfonyStyle->newLine(2);
     }
 }
