--- packages/monorepo-builder/packages/testing/src/Contract/ComposerJsonRepositoriesUpdaterInterface.php
@@ -1,17 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Symplify\MonorepoBuilder\Testing\Contract;
-
-use Symplify\SmartFileSystem\SmartFileInfo;
-
-interface ComposerJsonRepositoriesUpdaterInterface
-{
-    /**
-     * @param mixed[] $packageComposerJson
-     * @param string[] $packageNames
-     * @return mixed[]
-     */
-    public function decoratePackageComposerJson(array $packageComposerJson, array $packageNames, SmartFileInfo $rootComposerJsonFileInfo, ?bool $symlink): array;
-}
