--- packages/monorepo-builder/packages/testing/tests/ComposerJson/packages/package-three/composer.json
@@ -1,19 +1,0 @@
-{
-    "name": "example/package-three",
-    "repositories": [
-        {
-            "type": "path",
-            "url": "../../packages/package-one",
-            "options": {
-                "symlink": false
-            }
-        },
-        {
-            "type": "composer",
-            "url": "https://repo.packagist.com/acme-companies/"
-        },
-        {
-            "packagist.org": false
-        }
-    ]
-}
